resolved conflict in UserController and UserService
