+++ src/main/java/school/faang/user_service/service/SkillService.java
@@ -14,6 +14,7 @@ import school.faang.user_service.mapper.SkillCandidateMapper;
 import school.faang.user_service.mapper.SkillMapper;
 import school.faang.user_service.repository.SkillRepository;
 import school.faang.user_service.repository.recommendation.SkillOfferRepository;
+import school.faang.user_service.service.user.UserService;
 
 import java.util.List;
 import java.util.Optional;
