+++ src/main/java/school/faang/user_service/service/user/UserService.java
@@ -12,10 +12,13 @@ import school.faang.user_service.config.context.UserContext;
 import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.dto.user.UpdateUsersRankDto;
 import school.faang.user_service.entity.User;
+import school.faang.user_service.entity.UserSkillGuarantee;
+import school.faang.user_service.entity.recommendation.SkillOffer;
 import school.faang.user_service.exception.DataValidationException;
 import school.faang.user_service.mapper.UserMapper;
 import school.faang.user_service.entity.UserProfilePic;
 import school.faang.user_service.repository.UserRepository;
+import school.faang.user_service.repository.UserSkillGuaranteeRepository;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -32,10 +35,11 @@ public class UserService {
     private final static int BATCH_SIZE = 50;
     @PersistenceContext
     private final EntityManager entityManager;
-    private final UserRepository userRepository;
     private final UserMapper userMapper;
     private final UserContext userContext;
     private final AvatarService avatarService;
+    private final UserRepository userRepository;
+    private final UserSkillGuaranteeRepository userSkillGuaranteeRepository;
 
     public Optional<User> findById(long userId) {
         return userRepository.findById(userId);
@@ -110,5 +114,13 @@ public class UserService {
         return users.stream()
                 .map(userMapper::toDto)
                 .toList();
+    }
+
+    public UserSkillGuarantee addGuaranty(long userId, SkillOffer skillOffer) {
+        UserSkillGuarantee guarantee = UserSkillGuarantee.builder().user(
+                        userRepository.findById(userId).get()
+                ).guarantor(skillOffer.getRecommendation().getAuthor())
+                .build();
+        return userSkillGuaranteeRepository.save(guarantee);
     }
 }
